Add enabling logic
